Fixed ban from third parties (in follow and like lists); double-checked like and photo actions

+++ service/api/photo.go
@@ -10,6 +10,7 @@ import (
 )
 
 func (rt *_router) uploadPhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
+	// authenticate the user performing the action
 	user, code, err := rt.AuthenticateUserFromParameter("uname", r, ps)
 
 	if err != nil {
@@ -17,16 +18,17 @@ func (rt *_router) uploadPhoto(w http.ResponseWriter, r *http.Request, ps httpro
 		return
 	}
 
-	photo := PhotoDefault() // TODO: IN TEORIA DEVI PRENDERLO DAL REQUEST BODY PER PRENDERE ANCHE L'URL
+	photo := PhotoDefault() // TODO: va presa dal request body
 
 	photo.User = user
 
 	photo.Date = time.Now().Format("2006-01-02 15:04:05")
 
-	// TODO: MANCA DA CAPIRE COSA SIA L'URL
+	// TODO: capire cosa sia l'url
 
 	dbPhoto := photo.PhotoIntoDatabasePhoto()
 
+	// insert the photo into the database
 	err = rt.db.InsertPhoto(&dbPhoto)
 
 	if err != nil {
@@ -37,19 +39,22 @@ func (rt *_router) uploadPhoto(w http.ResponseWriter, r *http.Request, ps httpro
 	photo.Id = dbPhoto.Id
 
 	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusCreated)
+	w.WriteHeader(http.StatusCreated) // 201
 
+	// return the newly created photo
 	_ = json.NewEncoder(w).Encode(photo)
 }
 
 func (rt *_router) deletePhoto(w http.ResponseWriter, r *http.Request, ps httprouter.Params, ctx reqcontext.RequestContext) {
-	_, code, err := rt.AuthenticateUserFromParameter("uname", r, ps)
+	// authenticate the user performing the action
+	user, code, err := rt.AuthenticateUserFromParameter("uname", r, ps)
 
 	if err != nil {
 		http.Error(w, err.Error(), code)
 		return
 	}
 
+	// get the photo to be deleted from the resource parameter
 	photo, code, err := rt.GetPhotoFromParameter("photo_id", r, ps)
 
 	if err != nil {
@@ -57,6 +62,13 @@ func (rt *_router) deletePhoto(w http.ResponseWriter, r *http.Request, ps httpro
 		return
 	}
 
+	// check if the resource is consistent
+	if photo.User.Id != user.Id {
+		http.Error(w, ErrPageNotFound.Error(), http.StatusNotFound)
+		return
+	}
+
+	// remove the photo from the database
 	err = rt.db.DeletePhoto(photo.PhotoIntoDatabasePhoto())
 
 	if err != nil {
@@ -65,7 +77,8 @@ func (rt *_router) deletePhoto(w http.ResponseWriter, r *http.Request, ps httpro
 	}
 
 	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
+	w.WriteHeader(http.StatusOK) // 200
 
+	// return the removed photo
 	_ = json.NewEncoder(w).Encode(photo)
 }
